One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/planning/MerchSetup/merchdetail/merchdetail.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { MerchdetailComponent } from './merchdetail.component';

describe('Component: Merchdetail', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [MerchdetailComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([MerchdetailComponent],
      (component: MerchdetailComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(MerchdetailComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(MerchdetailComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-merchdetail></app-merchdetail>
  `,
  directives: [MerchdetailComponent]
})
class MerchdetailComponentTestController {
}

